requirements.txt
pytest
allure-pytest
Appium-Python-Client
